+++ memory-viz/bin/cli.js
@@ -4,6 +4,30 @@ const fs = require("fs");
 const path = require("path");
 const { draw } = require("memory-viz");
 const { program } = require("commander");
+const { json } = require("node:stream/consumers");
+
+function parseFilePath(input) {
+    if (input) {
+        return pathExists(input, `File`);
+    } else {
+        return undefined;
+    }
+}
+
+function parseOutputPath(input) {
+    return pathExists(input, `Output path`);
+}
+
+// helper function for parsing paths
+function pathExists(inputPath, errMsg) {
+    const fullPath = path.resolve(process.cwd(), inputPath);
+    if (!fs.existsSync(fullPath)) {
+        console.error(`Error: ${errMsg} ${fullPath} does not exist.`);
+        process.exit(1);
+    } else {
+        return fullPath;
+    }
+}
 
 function parseRoughjsConfig(input) {
     const pairs = input.split(",");
@@ -16,8 +40,14 @@ program
         "Command line interface for generating memory model diagrams with MemoryViz"
     )
     .argument(
-        "<filepath>",
-        "path to a file containing MemoryViz-compatible JSON"
+        "[filePath]",
+        "path to a file containing MemoryViz-compatible JSON",
+        parseFilePath
+    )
+    .option(
+        "--output <path>",
+        "writes generated SVG to specified path",
+        parseOutputPath
     )
     .option("--width <value>", "width of generated SVG", "1300")
     .option("--height <value>", "height of generated SVG")
@@ -29,43 +59,55 @@ program
     );
 
 program.parse();
+const filePath = program.processedArgs[0];
+const options = program.opts();
 
-const filePath = program.args[0];
-const absolutePath = path.resolve(process.cwd(), filePath);
+if (filePath) {
+    const fileContent = fs.readFileSync(filePath, "utf8");
 
-// Checks if absolutePath exists
-let fileContent;
-if (!fs.existsSync(absolutePath)) {
-    console.error(`Error: File ${absolutePath} does not exist.`);
-    process.exit(1);
+    let jsonContent;
+    try {
+        jsonContent = JSON.parse(fileContent);
+    } catch (err) {
+        console.error(`Error: Invalid JSON\n${err.message}.`);
+        process.exit(1);
+    }
+
+    runMemoryViz(jsonContent);
 } else {
-    fileContent = fs.readFileSync(absolutePath, "utf8");
+    json(process.stdin)
+        .then((jsonContent) => {
+            runMemoryViz(jsonContent);
+        })
+        .catch((err) => {
+            console.error(`Error: ${err.message}.`);
+            process.exit(1);
+        });
 }
 
-let data;
-try {
-    data = JSON.parse(fileContent);
-} catch (err) {
-    console.error(`Error: Invalid JSON\n${err.message}.`);
-    process.exit(1);
-}
+function runMemoryViz(jsonContent) {
+    let m;
+    try {
+        m = draw(jsonContent, true, {
+            width: options.width,
+            height: options.height,
+            roughjs_config: { options: options.roughjsConfig },
+        });
+    } catch (err) {
+        console.error(`Error: ${err.message}`);
+        process.exit(1);
+    }
 
-let m;
-try {
-    m = draw(data, true, {
-        width: program.opts().width,
-        height: program.opts().height,
-        roughjs_config: { options: program.opts().roughjsConfig },
-    });
-} catch (err) {
-    console.error(`Error: ${err.message}`);
-    process.exit(1);
-}
-
-const outputName = path.parse(filePath).name + ".svg";
-try {
-    m.save(outputName);
-} catch (err) {
-    console.error(`Error: ${err.message}`);
-    process.exit(1);
+    try {
+        if (options.output) {
+            const outputName = path.parse(options.output).name + ".svg";
+            const outputPath = path.join(options.output, outputName);
+            m.save(outputPath);
+        } else {
+            m.save();
+        }
+    } catch (err) {
+        console.error(`Error: ${err.message}`);
+        process.exit(1);
+    }
 }